fix some bug in backend (#18)

+++ src/github-profile/github-profile.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateGithubProfileDto } from './dto';
 
@@ -16,6 +16,27 @@ export class GithubProfileService {
   }
 
   async deleteGithubProfile(userId: number) {
+    const user = await this.prisma.user.findUnique({
+      where: {
+        id: userId,
+      },
+      include: {
+        githubProfile: true,
+      },
+    });
+
+    const projects = await this.prisma.project.findMany({
+      where: {
+        githubProfileId: user.githubProfile.id,
+      },
+    });
+
+    if (projects.length > 0) {
+      throw new ForbiddenException(
+        'Cannot delete Github Profile. It is associated with a project.',
+      );
+    }
+
     return this.prisma.githubProfile.delete({
       where: {
         userId,
